% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-adr.R
\name{adr_printer}
\alias{adr_printer}
\title{Normierter Verleger / Drucker (Alte Drucke)}
\usage{
adr_printer(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Normierter Verleger / Drucker (Alte Drucke)
}
