% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_filter_bibtype}
\alias{gbv_filter_bibtype}
\title{Einträge mit angegebener bibliographischer Gattung (bg)}
\usage{
gbv_filter_bibtype(items, bg)
}
\arguments{
\item{items}{Item data parsed from picaxml}

\item{bg}{Bibliographic type of item to be filtered}
}
\description{
Einträge mit angegebener bibliographischer Gattung (bg)
}
