% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_idn}
\alias{gbv_idn}
\title{Identifikationsnummer des Datensatzes (IDN)}
\usage{
gbv_idn(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Identifikationsnummer des Datensatzes (IDN)
}
