% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_publisher}
\alias{gbv_publisher}
\title{Veroeffentlichungsangabe / Verlagsname}
\usage{
gbv_publisher(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Item data parsed from picaxml}

\item{preserve}{Whether to preserve empty values}

\item{collapse}{Whether to collapse multiple values}
}
\description{
Veroeffentlichungsangabe / Verlagsname
}
