% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_url}
\alias{gbv_url}
\title{Elektronische Adresse und ergaenzende Angaben zum Zugriff (URL)}
\usage{
gbv_url(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Elektronische Adresse und ergaenzende Angaben zum Zugriff (URL)
}
