% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_urn}
\alias{gbv_urn}
\title{Elektronische Adresse und ergaenzende Angaben zum Zugriff (URL)}
\usage{
gbv_urn(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Elektronische Adresse und ergaenzende Angaben zum Zugriff (URL)
}
