% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_years}
\alias{gbv_years}
\title{Veroeffentlichungsangabe / Erscheinungsjahr in Originalschrift}
\usage{
gbv_years(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Veroeffentlichungsangabe / Erscheinungsjahr in Originalschrift
}
