% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{get_data}
\alias{get_data}
\title{Get pica items imported via catmandu and parsed as picaxml}
\usage{
get_data(fp)
}
\arguments{
\item{fp}{Path to json file with parsed pica data.}
}
\description{
Get pica items imported via catmandu and parsed as picaxml
}
