% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{get_data__pp}
\alias{get_data__pp}
\title{Get pica items imported via catmandu and parsed from picaplus-xml}
\usage{
get_data__pp(fp)
}
\arguments{
\item{fp}{Path to json file with parsed pica data.}
}
\description{
Get pica items imported via catmandu and parsed from picaplus-xml
}
