% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_sru}
\alias{import_sru}
\title{Import pica data via catmandu with given parameters}
\usage{
import_sru(
  base,
  query,
  recordSchema,
  parser = NULL,
  total = NULL,
  target = NULL,
  pretty = T
)
}
\arguments{
\item{base}{Url of SRU endpoint}

\item{query}{Query to be requested}

\item{recordSchema}{Record schema to be requested}

\item{parser}{Parser for record schema (optional)}

\item{total}{Total number of records to be requested}

\item{target}{Directory path for resulting json file}

\item{pretty}{Whether json data should be pretty printed}
}
\description{
Import pica data via catmandu with given parameters
}
