% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-adr.R
\name{printer_gnd}
\alias{printer_gnd}
\title{Normierter Verleger / Drucker (Alte Drucke)}
\usage{
printer_gnd(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Normierter Verleger / Drucker (Alte Drucke)
}
