% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tag}
\alias{tag}
\title{Extract tag indicated by id from given item}
\usage{
tag(item, id)
}
\arguments{
\item{item}{Item data parsed from picaxml}

\item{id}{Identifier of tag to be extracted}
}
\description{
Extract tag indicated by id from given item
}
