% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag__pp}
\alias{tag__pp}
\title{Extract tag indicated by id from given item}
\usage{
tag__pp(item, id)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}

\item{id}{Identifier of tag to be extracted}
}
\description{
Extract tag indicated by id from given item
}
