% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_subf__pp}
\alias{tag_subf__pp}
\title{Extract content of subf of tag indicated by id from given item}
\usage{
tag_subf__pp(item, id, subf)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}

\item{id}{Identifier of tag to be extracted}

\item{subf}{Subfield identifier of tag to be extracted}
}
\description{
Extract content of subf of tag indicated by id from given item
}
