% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_contenttype_gnd}
\alias{zdb_contenttype_gnd}
\title{Angaben zum Inhalt (GND-Nummer)}
\usage{
zdb_contenttype_gnd(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Angaben zum Inhalt (GND-Nummer)
}
