% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_contributor}
\alias{zdb_contributor}
\title{Person, Familie – Sonstige und Mitwirkende}
\usage{
zdb_contributor(item)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}
}
\description{
Person, Familie – Sonstige und Mitwirkende
}
