% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_contributor_gnd}
\alias{zdb_contributor_gnd}
\title{Person, Familie – Sonstige und Mitwirkende (GND)}
\usage{
zdb_contributor_gnd(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}

\item{preserve}{Whether to preserve empty values}

\item{collapse}{Whether to collapse multiple values}
}
\description{
Person, Familie – Sonstige und Mitwirkende (GND)
}
