% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_filter_bibtype}
\alias{zdb_filter_bibtype}
\title{Einträge mit angegebener bibliographischer Gattung (bg)}
\usage{
zdb_filter_bibtype(items, bg)
}
\arguments{
\item{bg}{Bibliographic type of item to be filtered}

\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Einträge mit angegebener bibliographischer Gattung (bg)
}
