% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_miscidx}
\alias{zdb_miscidx}
\title{Miscellaneous Identifiers}
\usage{
zdb_miscidx(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}
}
\description{
Miscellaneous Identifiers
}
