% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_organ}
\alias{zdb_organ}
\title{Körperschaft, Konferenz - 1. geistiger Schöpfer}
\usage{
zdb_organ(item)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}
}
\description{
Körperschaft, Konferenz - 1. geistiger Schöpfer
}
