% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_organ_name}
\alias{zdb_organ_name}
\title{Körperschaft, Konferenz - 1. geistiger Schöpfer (Name)}
\usage{
zdb_organ_name(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Körperschaft, Konferenz - 1. geistiger Schöpfer (Name)
}
