% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_organ_role}
\alias{zdb_organ_role}
\title{Körperschaft, Konferenz - 1. geistiger Schöpfer (Rolle)}
\usage{
zdb_organ_role(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}

\item{preserve}{Whether to preserve empty values}

\item{collapse}{Whether to collapse multiple values}
}
\description{
Körperschaft, Konferenz - 1. geistiger Schöpfer (Rolle)
}
