% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_person_role}
\alias{zdb_person_role}
\title{Person - Bevorzugter Name (Rolle)}
\usage{
zdb_person_role(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Person - Bevorzugter Name (Rolle)
}
