% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_place_prev}
\alias{zdb_place_prev}
\title{Frühere Verlagsorte und Verleger}
\usage{
zdb_place_prev(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Frühere Verlagsorte und Verleger
}
