% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_pubhistory_strict}
\alias{zdb_pubhistory_strict}
\title{maschinell verarbeitbarer Erscheinungsverlauf}
\usage{
zdb_pubhistory_strict(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
maschinell verarbeitbarer Erscheinungsverlauf
}
