% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_publisher_place}
\alias{zdb_publisher_place}
\title{Verlagsorte, Verlag, Verleger}
\usage{
zdb_publisher_place(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Verlagsorte, Verlag, Verleger
}
