% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-adr.R
\name{adr_printer_gnd}
\alias{adr_printer_gnd}
\title{Normierter Verleger / Drucker GND (Alte Drucke)}
\usage{
adr_printer_gnd(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Item data parsed from picaxml}

\item{preserve}{Whether to preserve empty values}

\item{collapse}{Whether to collapse multiple values}
}
\description{
Normierter Verleger / Drucker GND (Alte Drucke)
}
