% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-gbv.R
\name{gbv_maintitle}
\alias{gbv_maintitle}
\title{Haupttitel, Titelzusatz, Verantwortlichkeitsangabe}
\usage{
gbv_maintitle(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Haupttitel, Titelzusatz, Verantwortlichkeitsangabe
}
