% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_sru}
\alias{import_sru}
\title{Import pica data via catmandu with given parameters}
\usage{
import_sru(
  base,
  query,
  recordSchema,
  parser = NULL,
  total = NULL,
  target = NULL,
  pretty = T
)
}
\description{
Import pica data via catmandu with given parameters
}
