% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{tag_subf}
\alias{tag_subf}
\title{Extract content of subf of tag indicated by id from given item}
\usage{
tag_subf(item, id, subf)
}
\description{
Extract content of subf of tag indicated by id from given item
}
