% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_contributor_gnd}
\alias{zdb_contributor_gnd}
\title{Person, Familie – Sonstige und Mitwirkende (GND)}
\usage{
zdb_contributor_gnd(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Person, Familie – Sonstige und Mitwirkende (GND)
}
