% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_contributor_name}
\alias{zdb_contributor_name}
\title{Person, Familie – Sonstige und Mitwirkende (Name)}
\usage{
zdb_contributor_name(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Person, Familie – Sonstige und Mitwirkende (Name)
}
