% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_contributor_role}
\alias{zdb_contributor_role}
\title{Person, Familie – Sonstige und Mitwirkende (Rolle)}
\usage{
zdb_contributor_role(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Person, Familie – Sonstige und Mitwirkende (Rolle)
}
