% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_idn}
\alias{zdb_idn}
\title{Identifikationsnummer des Datensatzes (IDN)}
\usage{
zdb_idn(item)
}
\arguments{
\item{item}{Item data parsed from picaxml}
}
\description{
Identifikationsnummer des Datensatzes (IDN)
}
