% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_organ_role}
\alias{zdb_organ_role}
\title{Körperschaft, Konferenz - 1. geistiger Schöpfer (Rolle)}
\usage{
zdb_organ_role(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Körperschaft, Konferenz - 1. geistiger Schöpfer (Rolle)
}
