% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_person_name}
\alias{zdb_person_name}
\title{Person - Bevorzugter Name}
\usage{
zdb_person_name(item, collapse = FALSE)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}
}
\description{
Person - Bevorzugter Name
}
