% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_place_prev}
\alias{zdb_place_prev}
\title{Frühere Verlagsorte und Verleger}
\usage{
zdb_place_prev(item, preserve = FALSE, collapse = FALSE)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}

\item{preserve}{Whether to preserve empty values}

\item{collapse}{Whether to collapse multiple values}
}
\description{
Frühere Verlagsorte und Verleger
}
