% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_pubhistory_strict}
\alias{zdb_pubhistory_strict}
\title{Erscheinungsverlauf (maschinell verarbeitbar)}
\usage{
zdb_pubhistory_strict(item)
}
\arguments{
\item{item}{Pica item retrieved from ZDB}
}
\description{
Erscheinungsverlauf (maschinell verarbeitbar)
}
