% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-zdb.R
\name{zdb_publisher}
\alias{zdb_publisher}
\title{Verlagsorte, Verlag, Verleger}
\usage{
zdb_publisher(item)
}
\arguments{
\item{item}{Item data parsed from PicaPlus-XML}
}
\description{
Verlagsorte, Verlag, Verleger
}
